Make coldchain a plugin
