+++ backend/coldchain/Cargo.toml
@@ -11,6 +11,7 @@ chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
 tokio = { version = "1", features = ["macros"] }
+log = "0.4"
 
 [features]
 telegram-tests = ["service/telegram-tests"]
